requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.29
pdfplumber==0.6.1
nltk==3.7
nameparser==1.1.1
spacy==3.3.0
pyap==0.3.1
parsedatetime==2.6
boto3==1.22.4
